+++ modules/opening/src/main/OpeningPage.scala
@@ -33,16 +33,16 @@ case object NamePart:
   def from(moves: List[SanStr]): NamePartList = moves.map(Left.apply)
 
 case class ResultCounts(
-    white: Int,
-    draws: Int,
-    black: Int
+    white: Long,
+    draws: Long,
+    black: Long
 ):
-  lazy val sum: Int = white + draws + black
+  lazy val sum: Long = white + draws + black
 
-  def whitePercent                     = percentOf(white)
-  def drawsPercent                     = percentOf(draws)
-  def blackPercent                     = percentOf(black)
-  private def percentOf(v: Int): Float = (v.toFloat * 100 / sum)
+  def whitePercent                      = percentOf(white)
+  def drawsPercent                      = percentOf(draws)
+  def blackPercent                      = percentOf(black)
+  private def percentOf(v: Long): Float = (v.toFloat * 100 / sum)
 
 case class OpeningNext(
     san: String,
